+++ plugins/one_chat/lib/one_chat/settings/schema/layout.ex
@@ -4,6 +4,22 @@ defmodule OneChat.Settings.Schema.Layout do
   @primary_key {:id, :binary_id, autogenerate: true}
   @foreign_key_type :binary_id
 
+  @content_home_body_default """
+    # Welcome to InfinityOne
+
+    InfinityOne is your one-stop location of all your communication needs.
+
+    Click on one of the Rooms listed on the left navigation panel to get started.
+
+    Access InfinityOne in your browser, or download [The Desktop App](/apps) for an even better experience!
+
+    ## Helpful Links
+
+    * [Desktop Apps](/apps)
+    * [InfinityOne Overview](/pages)
+    * [Features](/features)
+    """
+
   schema "settings_layout" do
     field :display_roles, :boolean, default: true
     field :merge_private_groups, :boolean, default: true
@@ -14,8 +30,7 @@ defmodule OneChat.Settings.Schema.Layout do
       "'Segoe UI Emoji', 'Segoe UI Symbol', 'Meiryo UI'"
     field :content_home_title, :string, default: "Home"
     field :content_home_body, :string,
-      default: "Welcome to Ucx Chat <br> Go to APP SETTINGS -> Layout" <>
-      " to customize this intro."
+      default: @content_home_body_default
     field :content_side_nav_footer, :string,
       default: ~s(<img src="/images/logo.png" />)
   end
@@ -34,4 +49,6 @@ defmodule OneChat.Settings.Schema.Layout do
     struct
     |> cast(params, @fields)
   end
+
+  def content_home_body_default, do: @content_home_body_default
 end